% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{gen_data}
\alias{gen_data}
\title{Data generation function on the simulation part with given parameters}
\usage{
gen_data(n, beta, sigma, phi)
}
\arguments{
\item{n}{The number of observations}

\item{beta}{A 2-dimensional vector in the mean of conditional distribution of Y1 given X}

\item{sigma}{The standard deviation of conditional distribution of Y1 given X}

\item{phi}{A 3-dimensional vector included in the parameter of Y2 given X and Y1}
}
\value{
Data frame. Among variables of the data frame, y1_mis = y1 if delta1=1 and NA if delta1=0. Similarly, y2_mis = y2 if delta2=1 and NA if delta2=0.
}
\description{
Data generation function on the simulation part with given parameters
}
\examples{
gen_data(n=200, beta=c(1, 0.7), sigma=1, phi=c(-3, 0.5, 0.7))
}
